+++ ts/packages/knowProTest/src/index.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+export * from "./requests.js";
 export * from "./knowproContext.js";
 export * from "./knowproSearch.js";
 export * from "./memoryCommands.js";
